+++ server/src/main/java/com/exchangediary/diary/domain/DiaryRepository.java
@@ -1,5 +1,6 @@
-package com.exchangediary.diary.domain.entity;
+package com.exchangediary.diary.domain;
 
+import com.exchangediary.diary.domain.entity.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiaryRepository extends JpaRepository<Diary, Long> {
