+++ server/src/main/java/com/exchangediary/diary/service/DiaryCommandService.java
@@ -2,7 +2,7 @@ package com.exchangediary.diary.service;
 
 import com.exchangediary.diary.domain.UploadImageRepository;
 import com.exchangediary.diary.domain.entity.Diary;
-import com.exchangediary.diary.domain.entity.DiaryRepository;
+import com.exchangediary.diary.domain.DiaryRepository;
 import com.exchangediary.diary.domain.entity.PublicationStatus;
 import com.exchangediary.diary.domain.entity.UploadImage;
 import com.exchangediary.diary.ui.dto.request.DiaryRequest;
