+++ server/src/main/java/com/exchangediary/diary/ui/DiaryController.java
@@ -1,7 +1,9 @@
-package com.exchangediary.diary.ui.dto.request;
+package com.exchangediary.diary.ui;
 
 import com.exchangediary.diary.domain.entity.Diary;
 import com.exchangediary.diary.service.DiaryCommandService;
+import com.exchangediary.diary.ui.dto.request.DiaryRequest;
+import com.exchangediary.diary.ui.dto.request.UploadImageRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
